requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
django-heroku
